fix: se repararon errores de atributos, nombres y logica menor en pro de optimizar el servicio, esto se hizo probando junto a postman para detectar errores

+++ msvc_payment/src/main/java/com/GameHub/models/Payment.java
@@ -28,7 +28,7 @@ public class Payment {
     private Long ordenId;
 
     @NotNull(message = "El monto es obligatorio")
-    private double monto;
+    private Long monto;
 
     @NotBlank(message = "El método de pago es obligatorio")
     private String metodo;
