+++ msvc_payment/src/main/java/com/GameHub/models/dtos/PaymentSaveDTO.java
@@ -15,7 +15,7 @@ public class PaymentSaveDTO {
     @NotNull(message = "El ID de la orden es obligatorio")
     private Long ordenId;
     @NotNull(message = "El monto es obligatorio")
-    private double monto;
+    private Long monto;
     @NotBlank(message = "El método de pago es obligatorio")
     private String metodo;
     private String codigoTransaccion;
